# backend/src/Domain/DTOs/CreateProductDto.cs
using Domain.Enums;
using System.ComponentModel.DataAnnotations;

namespace Domain.DTOs;

public class CreateProductDto
{
    [Required]
    [StringLength(200)]
    public string Name { get; set; } = string.Empty;

    [Required]
    [StringLength(2000)]
    public string Description { get; set; } = string.Empty;

    [Required]
    [Range(0.01, double.MaxValue, ErrorMessage = "Price must be greater than 0")]
    public decimal Price { get; set; }

    [Required]
    public int CategoryId { get; set; }
}