# backend/src/Domain/DTOs/ProductDto.cs
namespace Domain.DTOs;

public class ProductDto
{
    public int Id { get; set; }
    public string Name { get; set; } = string.Empty;
    public string Description { get; set; } = string.Empty;
    public decimal Price { get; set; }
    public string Image { get; set; } = string.Empty;

    public CategoryDto Category { get; set; }
}